extract extendable functions as mixin for AudioSelector to extend
